chore: move http.Handler to http.go (#16775)
